move operator to sync package
