+++ pkg/sync/workdir/aur_source.go
@@ -1,4 +1,4 @@
-package main
+package workdir
 
 import (
 	"context"
