+++ pkg/sync/workdir/preparer_test.go
@@ -1,7 +1,7 @@
 //go:build !integration
 // +build !integration
 
-package main
+package workdir
 
 import (
 	"testing"
